Add a new daemon for stats monitoring
Summary:
ksysguardd, whilst good, has a few problems

The code is a bit archaic, it relies on a polling API, which is overhead
for infrequently changed values or where setting up a monitor has a big
overhead.

It also moves the problem of translations into the daemon, allowing for
better extensibility without requiring client side changes.

The daemon is based around a typical OO model. Plugins have lists of
objects, those objects have properties using common Qt patterns. A
property also has various metadata.

For full compatibility ksgrd is wrapped and the plan is to land with the
bridge, then slowly land patches that use the new API natively.

An nvidia plugin is also added to show the API being used in another
format.

This is all consumed by the new API posted in D28141

Test Plan:
Unit test
Used with the new library to create a new suite of applet (upcoming patch)
Used in a ported ksysguard

Reviewers: #plasma, mart, ngraham, ahiemstra

Reviewed By: #plasma, mart, ngraham, ahiemstra

Subscribers: ahiemstra, ivan, mart, zzag, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D28333