Add a network backend that uses rtnetlink
Queries information about current network devices using the rtnetlink kernel
interface. This exposed all the info except a pretty name for the current
connection and wireless strength. We can use this as a fallback on Linux if
NetworkManager is not available.
Because this is our fallback on Linux libnl is now required (from recommended).
BUG:425992