NetworkManager: Do not remove devices when their active connection changes
In certain cases, NetworkManager can go on a rampage sending active
connection change signals in very rapid succession. This results in
memory usage ballooning because lambda connections can not be unique.

Rather than destroying and recreating the NetworkManagerDevice objects
on active connection change, this changes things so that we only remove
them from the SensorContainer but keep the actual objects around. This
allows NetworkManagerDevice to track its active connection state,
removing the need for constant reconnects.

BUG: 430003