Check if we could open a file
In order to read the file it not only does need to exist but we also have to open it.
This broke some systems where a file existed but was not readable with current permission.
If we check if we could open a file we can save the check if the file exists.