Add more network sensor
Starting to use plasma system monitor on a daily basis, I noticed the lack of some sensors, so I decided to implement the ones I needed the most. In this first pr I focused on some network sensors: IPv4/6 gateway, subnetmask/prefix length and dns servers. The work should be completed, but I still left a FIXME for the case of the dns sensors when we are not using network manager to obtaining the information: the point is that I don't even think there exists a reliable way to retrieve the currently used dns in general, so those sensors return something only when networkmanager is available.