New CPU cumulative sensors
This commit adds a few cumulative sensor for the CPU: the max/min/average value
of the clock frequencies and temperatures of the various cores/CPUs. The reason
is quite simple: modern CPUs have more and more cores and all of them have their
own sensors; with such large number of sensors, a summary of those is needed if
you want to know the state of your processor at a glance.