daemon: Replace KDBusService with manual implementation of replace
KDBusService does not allow us to specify which service to register or
at what path. It also includes a bunch of things like window activation
that is rather useless for something that doesn't have a window. In
fact, we really only use it for it's uniqueness and replace behaviour.
Since we now need control over what service name and object paths to
register, drop KDBusService and use QDBusConnectionInterface's replace
functionality instead.