gpu/nvidia: Store memory values as bytes
The "all gpus" aggregate sensors use bytes as unit. However, the NVidia
sensors are set to use megabytes. This results in the aggregate sensor
using an incorrect and too low value for the "all gpu" sensors. So
instead, make sure to store the value for NVidia sensors as bytes so
aggregation is correct.

BUG: 477687