Create a GPU backend for FreeBSD
This commit create a FreeBSD GPU sensor backend. In addition to that, an NVidia GPU backend is also created.
It's very similar to the Linux backends. The main difference is that device discovery is done through devinfo library of FreeBSD, instead of udev.

I've tested on my FreeBSD machine, and it works for my NVidia GTX 1070.