Fix buffer memory reporting
We can't use 'uint64_t' here, because it's mapped to 'ulong' Qt type,
which is not registered in Qt DBus type system. On the other hand,
'unsigned long long' is okay, because it maps to 'qulonglong', which
maps to DBus type 'uint64', which is what we want.
See https://doc.qt.io/qt-6/qdbustypesystem.html for details.

BUG: 498373