plugins/cpu: Use physical ID to determine number of CPUs
The ID from the cpu info has no relation to the number of physical CPUs.
For that, we need to check the physical ID of the CPU. So use that if
it's available. If it's not available, we end up defaulting to 1 CPU
which is probably what we want anyway.

BUG: 496847